# docs/api/v1.0.0/ai-service-openapi.yaml
openapi: 3.0.3
info:
  title: AI Conversation Summarization API
  description: |
    RESTful API for AI-powered conversation summarization services.
    
    This API provides intelligent summarization capabilities for conversation data,
    allowing clients to generate concise summaries of chat conversations, meeting
    transcripts, and other conversational content.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com
    description: Production server
  - url: https://staging-api.example.com
    description: Staging server
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
paths:
  /api/v1/ai/summarize:
    post:
      summary: Generate conversation summary
      description: |
        Generates an AI-powered summary of a conversation based on the provided
        conversation ID. The summary includes key topics, important decisions,
        action items, and relevant context from the conversation history.
        
        The response includes metadata about when the summary was generated and
        when it expires from cache.
      operationId: summarizeConversation
      tags:
        - AI Summarization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SummarizationRequest'
            examples:
              basic_request:
                summary: Basic summarization request
                value:
                  conversation_id: "conv_123456789"
              with_limit:
                summary: Request with message limit
                value:
                  conversation_id: "conv_123456789"
                  limit: 50
      responses:
        '200':
          description: Successfully generated conversation summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SummarizationResponse'
              examples:
                success_response:
                  summary: Successful summarization
                  value:
                    summary: |
                      ## Key Topics Discussed
                      - Product roadmap planning for Q2 2024
                      - Budget allocation for new feature development
                      - Team restructuring and hiring plans
                      
                      ## Decisions Made
                      - Approved $100k budget for mobile app redesign
                      - Decided to postpone internationalization until Q3
                      
                      ## Action Items
                      - Sarah to draft technical specifications by Friday
                      - Mike to schedule follow-up with design team
                      - Team lead to present findings at next board meeting
                    generated_at: "2024-01-15T10:30:00Z"
                    expires_at: "2024-01-15T22:30:00Z"
        '400':
          description: Bad request - Invalid input parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid_conversation_id:
                  summary: Invalid conversation ID format
                  value:
                    error:
                      code: "INVALID_CONVERSATION_ID"
                      message: "The provided conversation ID format is invalid"
                      details: "conversation_id must be a string starting with 'conv_'"
                    request_id: "req_789012345"
                    timestamp: "2024-01-15T10:30:00Z"
                invalid_limit:
                  summary: Invalid limit parameter
                  value:
                    error:
                      code: "INVALID_LIMIT"
                      message: "The limit parameter must be between 1 and 1000"
                      details: "Provided limit: 1500. Maximum allowed: 1000"
                    request_id: "req_789012346"
                    timestamp: "2024-01-15T10:30:00Z"
        '401':
          description: Unauthorized - Invalid or missing authentication token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missing_token:
                  summary: Missing authentication token
                  value:
                    error:
                      code: "MISSING_AUTH_TOKEN"
                      message: "Authentication token is required"
                      details: "Please provide a valid Bearer token in the Authorization header"
                    request_id: "req_789012347"
                    timestamp: "2024-01-15T10:30:00Z"
                invalid_token:
                  summary: Invalid authentication token
                  value:
                    error:
                      code: "INVALID_AUTH_TOKEN"
                      message: "The provided authentication token is invalid or expired"
                      details: "Token validation failed"
                    request_id: "req_789012348"
                    timestamp: "2024-01-15T10:30:00Z"
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                insufficient_permissions:
                  summary: Insufficient permissions
                  value:
                    error:
                      code: "INSUFFICIENT_PERMISSIONS"
                      message: "You don't have permission to access this conversation"
                      details: "Access to conversation 'conv_123456789' requires SUMMARIZE permission"
                    request_id: "req_789012349"
                    timestamp: "2024-01-15T10:30:00Z"
        '404':
          description: Not found - Conversation does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                conversation_not_found:
                  summary: Conversation not found
                  value:
                    error:
                      code: "CONVERSATION_NOT_FOUND"
                      message: "The specified conversation could not be found"
                      details: "No conversation exists with ID 'conv_123456789'"
                    request_id: "req_789012350"
                    timestamp: "2024-01-15T10:30:00Z"
        '429':
          description: Too many requests - Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                rate_limit_exceeded:
                  summary: Rate limit exceeded
                  value:
                    error:
                      code: "RATE_LIMIT_EXCEEDED"
                      message: "Rate limit exceeded. Please try again later"
                      details: "Maximum 100 requests per hour allowed. Reset at 2024-01-15T11:00:00Z"
                    request_id: "req_789012351"
                    timestamp: "2024-01-15T10:30:00Z"
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying
              schema:
                type: integer
                example: 3600
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internal_error:
                  summary: Internal server error
                  value:
                    error:
                      code: "INTERNAL_SERVER_ERROR"
                      message: "An unexpected error occurred while processing your request"
                      details: "Please try again later or contact support if the problem persists"
                    request_id: "req_789012352"
                    timestamp: "2024-01-15T10:30:00Z"
        '503':
          description: Service unavailable - AI service is temporarily unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                service_unavailable:
                  summary: AI service unavailable
                  value:
                    error:
                      code: "AI_SERVICE_UNAVAILABLE"
                      message: "AI summarization service is temporarily unavailable"
                      details: "The service is undergoing maintenance. Expected availability: 2024-01-15T12:00:00Z"
                    request_id: "req_789012353"
                    timestamp: "2024-01-15T10:30:00Z"
          headers:
            Retry-After:
              description: Number of seconds to wait before retrying
              schema:
                type: integer
                example: 7200
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT-based authentication. Include the token in the Authorization header
        as: `Authorization: Bearer <token>`
        
        The JWT token should contain the following claims:
        - `sub`: User ID
        - `iat`: Issued at timestamp
        - `exp`: Expiration timestamp
        - `scope`: Permissions (e.g., "ai:summarize")
  schemas:
    SummarizationRequest:
      type: object
      required:
        - conversation_id
      properties:
        conversation_id:
          type: string
          pattern: '^conv_[a-zA-Z0-9_-]+$'
          description: |
            Unique identifier for the conversation to be summarized.
            Must start with 'conv_' prefix followed by alphanumeric characters,
            underscores, or hyphens.
          example: "conv_123456789"
          minLength: 5
          maxLength: 100
        limit:
          type: integer
          description: |
            Optional limit on the number of messages to include in the summarization.
            If not provided, all available messages will be processed.
            Maximum value is 1000 messages.
          minimum: 1
          maximum: 1000
          default: 100
          example: 50
      additionalProperties: false
    SummarizationResponse:
      type: object
      required:
        - summary
        - generated_at
        - expires_at
      properties:
        summary:
          type: string
          description: |
            AI-generated summary of the conversation in Markdown format.
            Includes key topics, decisions made, action items, and important context.
            The summary is structured with headings and bullet points for readability.
          example: |
            ## Key Topics Discussed
            - Product roadmap planning for Q2 2024
            - Budget allocation for new feature development
            
            ## Decisions Made
            - Approved $100k budget for mobile app redesign
            
            ## Action Items
            - Sarah to draft technical specifications by Friday
          minLength: 1
          maxLength: 5000
        generated_at:
          type: string
          format: date-time
          description: |
            ISO 8601 timestamp indicating when the summary was generated.
            This timestamp is in UTC timezone.
          example: "2024-01-15T10:30:00Z"
        expires_at:
          type: string
          format: date-time
          description: |
            ISO 8601 timestamp indicating when this summary expires from cache.
            After this time, a new request will generate a fresh summary.
            Typically set to 12 hours after generation.
          example: "2024-01-15T22:30:00Z"
      additionalProperties: false
    ErrorResponse:
      type: object
      required:
        - error
        - request_id
        - timestamp
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetails'
        request_id:
          type: string
          description: |
            Unique identifier for this request, useful for debugging and support.
            Format: req_ followed by alphanumeric characters.
          pattern: '^req_[a-zA-Z0-9]+$'
          example: "req_789012345"
        timestamp:
          type: string
          format: date-time
          description: |
            ISO 8601 timestamp when the error occurred (UTC timezone).
          example: "2024-01-15T10:30:00Z"
      additionalProperties: false
    ErrorDetails:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: |
            Machine-readable error code that categorizes the type of error.
            Used for programmatic error handling.
          enum:
            - "INVALID_CONVERSATION_ID"
            - "INVALID_LIMIT"
            - "MISSING_AUTH_TOKEN"
            - "INVALID_AUTH_TOKEN"
            - "INSUFFICIENT_PERMISSIONS"
            - "CONVERSATION_NOT_FOUND"
            - "RATE_LIMIT_EXCEEDED"
            - "INTERNAL_SERVER_ERROR"
            - "AI_SERVICE_UNAVAILABLE"
          example: "INVALID_CONVERSATION_ID"
        message:
          type: string
          description: |
            Human-readable error message describing what went wrong.
            Safe to display to end users.
          example: "The provided conversation ID format is invalid"
          minLength: 1
          maxLength: 200
        details:
          type: string
          description: |
            Optional additional details about the error.
            May contain technical information useful for debugging.
          example: "conversation_id must be a string starting with 'conv_'"
          maxLength: 500
      additionalProperties: false
...